requirements.txt
fastapi
uvicorn
pydantic
scikit-learn
pandas
numpy
matplotlib
tensorflow
requests
streamlit
flask
python-dotenv
pytest
google-cloud-vision
azure-storage-blob
google-cloud-speech
google-cloud-core
grpcio
python-multipart
sounddevice
httpx
azure-cosmos
opencv-contrib-python-headless

